Collapse subflows into parents
