+++ graphrag/index/workflows/v1/create_final_covariates.py
@@ -3,11 +3,22 @@
 
 """A module containing build_steps method definition."""
 
+from typing import Any, cast
+
 from datashaper import (
     AsyncType,
+    Table,
+    VerbCallbacks,
+    verb,
 )
+from datashaper.table_store.types import VerbResult, create_verb_result
 
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.config.workflow import PipelineWorkflowConfig, PipelineWorkflowStep
+from graphrag.index.flows.create_final_covariates import (
+    create_final_covariates as create_final_covariates_flow,
+)
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 workflow_name = "create_final_covariates"
 
@@ -38,3 +49,32 @@ def build_steps(
             "input": {"source": "workflow:create_base_text_units"},
         },
     ]
+
+
+@verb(name="create_final_covariates", treats_input_tables_as_immutable=True)
+async def create_final_covariates(
+    callbacks: VerbCallbacks,
+    cache: PipelineCache,
+    runtime_storage: PipelineStorage,
+    covariate_type: str,
+    extraction_strategy: dict[str, Any] | None,
+    async_mode: AsyncType = AsyncType.AsyncIO,
+    entity_types: list[str] | None = None,
+    num_threads: int = 4,
+    **_kwargs: dict,
+) -> VerbResult:
+    """All the steps to extract and format covariates."""
+    text_units = await runtime_storage.get("base_text_units")
+
+    output = await create_final_covariates_flow(
+        text_units,
+        callbacks,
+        cache,
+        covariate_type,
+        extraction_strategy,
+        async_mode=async_mode,
+        entity_types=entity_types,
+        num_threads=num_threads,
+    )
+
+    return create_verb_result(cast("Table", output))
