+++ graphrag/index/workflows/v1/create_final_documents.py
@@ -3,9 +3,25 @@
 
 """A module containing build_steps method definition."""
 
-from datashaper import DEFAULT_INPUT_NAME
+from typing import TYPE_CHECKING, cast
+
+from datashaper import (
+    DEFAULT_INPUT_NAME,
+    Table,
+    VerbInput,
+    verb,
+)
+from datashaper.table_store.types import VerbResult, create_verb_result
 
 from graphrag.index.config.workflow import PipelineWorkflowConfig, PipelineWorkflowStep
+from graphrag.index.flows.create_final_documents import (
+    create_final_documents as create_final_documents_flow,
+)
+from graphrag.storage.pipeline_storage import PipelineStorage
+
+if TYPE_CHECKING:
+    import pandas as pd
+
 
 workflow_name = "create_final_documents"
 
@@ -30,3 +46,22 @@ def build_steps(
             },
         },
     ]
+
+
+@verb(
+    name="create_final_documents",
+    treats_input_tables_as_immutable=True,
+)
+async def create_final_documents(
+    input: VerbInput,
+    runtime_storage: PipelineStorage,
+    document_attribute_columns: list[str] | None = None,
+    **_kwargs: dict,
+) -> VerbResult:
+    """All the steps to transform final documents."""
+    source = cast("pd.DataFrame", input.get_input())
+    text_units = await runtime_storage.get("base_text_units")
+
+    output = create_final_documents_flow(source, text_units, document_attribute_columns)
+
+    return create_verb_result(cast("Table", output))
