+++ graphrag/index/workflows/v1/create_final_nodes.py
@@ -3,7 +3,20 @@
 
 """A module containing build_steps method definition."""
 
+from typing import Any, cast
+
+from datashaper import (
+    Table,
+    VerbCallbacks,
+    verb,
+)
+from datashaper.table_store.types import VerbResult, create_verb_result
+
 from graphrag.index.config.workflow import PipelineWorkflowConfig, PipelineWorkflowStep
+from graphrag.index.flows.create_final_nodes import (
+    create_final_nodes as create_final_nodes_flow,
+)
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 workflow_name = "create_final_nodes"
 
@@ -59,3 +72,33 @@ def build_steps(
             },
         },
     ]
+
+
+@verb(name="create_final_nodes", treats_input_tables_as_immutable=True)
+async def create_final_nodes(
+    callbacks: VerbCallbacks,
+    runtime_storage: PipelineStorage,
+    layout_strategy: dict[str, Any],
+    embedding_strategy: dict[str, Any] | None = None,
+    **_kwargs: dict,
+) -> VerbResult:
+    """All the steps to transform final nodes."""
+    base_entity_nodes = await runtime_storage.get("base_entity_nodes")
+    base_relationship_edges = await runtime_storage.get("base_relationship_edges")
+    base_communities = await runtime_storage.get("base_communities")
+
+    output = create_final_nodes_flow(
+        base_entity_nodes,
+        base_relationship_edges,
+        base_communities,
+        callbacks,
+        layout_strategy,
+        embedding_strategy=embedding_strategy,
+    )
+
+    return create_verb_result(
+        cast(
+            "Table",
+            output,
+        )
+    )
