+++ graphrag/index/workflows/v1/extract_graph.py
@@ -3,11 +3,23 @@
 
 """A module containing build_steps method definition."""
 
+from typing import Any, cast
+
+import pandas as pd
 from datashaper import (
     AsyncType,
+    Table,
+    VerbCallbacks,
+    verb,
 )
+from datashaper.table_store.types import VerbResult, create_verb_result
 
+from graphrag.cache.pipeline_cache import PipelineCache
 from graphrag.index.config.workflow import PipelineWorkflowConfig, PipelineWorkflowStep
+from graphrag.index.flows.extract_graph import (
+    extract_graph as extract_graph_flow,
+)
+from graphrag.storage.pipeline_storage import PipelineStorage
 
 workflow_name = "extract_graph"
 
@@ -50,3 +62,46 @@ def build_steps(
             "input": ({"source": "workflow:create_base_text_units"}),
         },
     ]
+
+
+@verb(
+    name="extract_graph",
+    treats_input_tables_as_immutable=True,
+)
+async def extract_graph(
+    callbacks: VerbCallbacks,
+    cache: PipelineCache,
+    storage: PipelineStorage,
+    runtime_storage: PipelineStorage,
+    extraction_strategy: dict[str, Any] | None,
+    extraction_num_threads: int = 4,
+    extraction_async_mode: AsyncType = AsyncType.AsyncIO,
+    entity_types: list[str] | None = None,
+    summarization_strategy: dict[str, Any] | None = None,
+    summarization_num_threads: int = 4,
+    snapshot_graphml_enabled: bool = False,
+    snapshot_transient_enabled: bool = False,
+    **_kwargs: dict,
+) -> VerbResult:
+    """All the steps to create the base entity graph."""
+    text_units = await runtime_storage.get("base_text_units")
+
+    base_entity_nodes, base_relationship_edges = await extract_graph_flow(
+        text_units,
+        callbacks,
+        cache,
+        storage,
+        extraction_strategy=extraction_strategy,
+        extraction_num_threads=extraction_num_threads,
+        extraction_async_mode=extraction_async_mode,
+        entity_types=entity_types,
+        summarization_strategy=summarization_strategy,
+        summarization_num_threads=summarization_num_threads,
+        snapshot_graphml_enabled=snapshot_graphml_enabled,
+        snapshot_transient_enabled=snapshot_transient_enabled,
+    )
+
+    await runtime_storage.set("base_entity_nodes", base_entity_nodes)
+    await runtime_storage.set("base_relationship_edges", base_relationship_edges)
+
+    return create_verb_result(cast("Table", pd.DataFrame()))
